LightFileReaderAndWriterForConfigModel: Ground a usage for FileReader and FileWriter in testcases
Steps of the refactor:
- SourceWriter and DestinationReader related wrapper functions
(read_log(), writer_log()) has been moved to FileReader and FileWriter
classes with their logic
- Reference for SourceWriter and DestinationReader classes has
been removed from FileSource, FileDestination and replaced in SNMPtrapd
- Reference of using direct message reading from file in ExampleDestination
has been removed
- After that SourceWriter and DestinationReader classes has been deleted
