+++ tests/python_functional/src/syslog_ng_config/statements/destinations/example_destination.py
@@ -23,8 +23,6 @@
 from pathlib2 import Path
 
 import src.testcase_parameters.testcase_parameters as tc_parameters
-from src.message_reader.message_reader import MessageReader
-from src.message_reader.single_line_parser import SingleLineParser
 from src.syslog_ng_config.statements.destinations.destination_driver import DestinationDriver
 
 
@@ -34,13 +32,5 @@ class ExampleDestination(DestinationDriver):
         self.path = Path(tc_parameters.WORKING_DIR, filename)
         super(ExampleDestination, self).__init__(None, dict({"filename": self.path.resolve()}, **options))
 
-    def wait_file_content(self, content):
-        with self.path.open() as f:
-            message_reader = MessageReader(f.readline, SingleLineParser())
-
-            while True:
-                msg = message_reader.pop_messages(1)[0]
-                if content in msg:
-                    return True
-
-            return False
+    def get_path(self):
+        return self.path
