+++ tests/python_functional/src/syslog_ng_config/statements/destinations/file_destination.py
@@ -23,9 +23,7 @@
 from pathlib2 import Path
 
 import src.testcase_parameters.testcase_parameters as tc_parameters
-from src.driver_io.file.file_reader import FileReader
 from src.syslog_ng_config.statements.destinations.destination_driver import DestinationDriver
-from src.syslog_ng_config.statements.destinations.destination_reader import DestinationReader
 
 
 class FileDestination(DestinationDriver):
@@ -33,13 +31,6 @@ class FileDestination(DestinationDriver):
         self.driver_name = "file"
         self.path = Path(tc_parameters.WORKING_DIR, file_name)
         super(FileDestination, self).__init__([self.path], options)
-        self.destination_reader = DestinationReader(FileReader)
 
     def get_path(self):
         return self.path
-
-    def read_log(self):
-        return self.destination_reader.read_logs(self.get_path(), counter=1)[0]
-
-    def read_logs(self, counter):
-        return self.destination_reader.read_logs(self.get_path(), counter=counter)
